requirements.txt
torchvision
Pillow
scipy
matplotlib
torch
